backend/requirements.txt
fastapi
uvicorn
pandas
openai
requests
pyyaml
python-dotenv 